+++ celeste/src/main.rs
@@ -3,6 +3,7 @@
 #![feature(panic_info_message)]
 #![feature(async_closure)]
 #![feature(trait_alias)]
+#![feature(exit_status_error)]
 
 pub mod about;
 pub mod entities;